Better page composition and quotes
Added the use of Jinja 'extends' so that web pages can share common elements.  This is currently used to have the time and battery percentage on every page using a 'base' page that other pages extend.

In the prior iteration the word jumble used dictionary words which went poorly as there are a lot of oddball words and proper words in the dictionary.  I replaced it with a list of 3000 of the most common US English words.   I removed from that list words that were two small and hyphenated words.

Added a simple 'Quote' page.